separated AWS CLI into subfolder, updated some descriptions of AWS classes
